Rename module to crypto_pwhash
